Use vscode APIs for file access
