+++ agent/src/vscode-shim.ts
@@ -1,6 +1,5 @@
 /* eslint-disable @typescript-eslint/no-empty-function */
 import { execSync } from 'child_process'
-import * as fspromises from 'fs/promises'
 import path from 'path'
 
 import type * as vscode from 'vscode'
@@ -37,6 +36,7 @@ import {
     UIKind,
     Uri,
     ViewColumn,
+    workspaceFs,
 } from '../../vscode/src/testutils/mocks'
 
 import type { Agent } from './agent'
@@ -152,70 +152,6 @@ export function setWorkspaceDocuments(newWorkspaceDocuments: WorkspaceDocuments)
 }
 
 export const workspaceFolders: vscode.WorkspaceFolder[] = []
-const fs: typeof vscode.workspace.fs = {
-    stat: async uri => {
-        const stat = await fspromises.stat(uri.fsPath)
-        const type = stat.isFile()
-            ? FileType.File
-            : stat.isDirectory()
-            ? FileType.Directory
-            : stat.isSymbolicLink()
-            ? FileType.SymbolicLink
-            : FileType.Unknown
-
-        return {
-            type,
-            ctime: stat.ctimeMs,
-            mtime: stat.mtimeMs,
-            size: stat.size,
-        }
-    },
-    readDirectory: async uri => {
-        const entries = await fspromises.readdir(uri.fsPath, { withFileTypes: true })
-
-        return entries.map(entry => {
-            const type = entry.isFile()
-                ? FileType.File
-                : entry.isDirectory()
-                ? FileType.Directory
-                : entry.isSymbolicLink()
-                ? FileType.SymbolicLink
-                : FileType.Unknown
-
-            return [entry.name, type]
-        })
-    },
-    createDirectory: async uri => {
-        await fspromises.mkdir(uri.fsPath, { recursive: true })
-    },
-    readFile: async uri => {
-        const content = await fspromises.readFile(uri.fsPath)
-        return new Uint8Array(content.buffer)
-    },
-    writeFile: async (uri, content) => {
-        await fspromises.writeFile(uri.fsPath, content)
-    },
-    delete: async (uri, options) => {
-        await fspromises.rm(uri.fsPath, { recursive: options?.recursive ?? false })
-    },
-    rename: async (source, target, options) => {
-        if (options?.overwrite ?? false) {
-            await fspromises.unlink(target.fsPath)
-        }
-        await fspromises.link(source.fsPath, target.fsPath)
-        await fspromises.unlink(source.fsPath)
-    },
-    copy: async (source, target, options) => {
-        const mode = options?.overwrite ? 0 : fspromises.constants.COPYFILE_EXCL
-        await fspromises.copyFile(source.fsPath, target.fsPath, mode)
-    },
-    isWritableFileSystem: scheme => {
-        if (scheme === 'file') {
-            return true
-        }
-        return false
-    },
-}
 
 // vscode.workspace.onDidChangeConfiguration
 const _workspace: typeof vscode.workspace = {
@@ -256,7 +192,7 @@ const _workspace: typeof vscode.workspace = {
             if (token?.isCancellationRequested) {
                 return
             }
-            const files = await fs.readDirectory(dir)
+            const files = await workspaceFs.readDirectory(dir)
             for (const [name, fileType] of files) {
                 const uri = Uri.file(path.join(dir.fsPath, name))
                 const relativePath = path.relative(workspaceRoot.fsPath, uri.fsPath)
@@ -277,7 +213,7 @@ const _workspace: typeof vscode.workspace = {
         await Promise.all(
             workspaceFolders.map(async folder => {
                 try {
-                    const stat = await fs.stat(folder.uri)
+                    const stat = await workspaceFs.stat(folder.uri)
                     if (stat.type.valueOf() === FileType.Directory.valueOf()) {
                         await loop(folder.uri, folder.uri)
                     }
@@ -350,7 +286,7 @@ const _workspace: typeof vscode.workspace = {
         }
         return configuration
     },
-    fs,
+    fs: workspaceFs,
 }
 
 export const workspace = _workspace
